+++ plugin/setup.php
@@ -48,6 +48,8 @@ function plugin_init_iservice(): void
         $PLUGIN_HOOKS['add_javascript']['iservice'] = "js/import.js";
 
         $PLUGIN_HOOKS['redefine_menus']['iservice'] = 'plugin_iservice_redefine_menus';
+
+        PluginIserviceConfig::handleConfigValues();
     }
 }